% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR.R
\name{DR}
\alias{DR}
\title{Douglas Rachford algorithm}
\usage{
DR(Delta_1, Delta_2, Delta_3, S_aux, rho, UT, LT)
}
\arguments{
\item{Delta_1}{Previous value to be updated.}

\item{Delta_2}{Previous value to be updated.}

\item{Delta_3}{Previous value to be updated.}

\item{S_aux}{Upper triangle of the matrix diag(xi) S diag(xi)}

\item{rho}{Penalty parameter.}

\item{UT}{Matrix indicating upper triangle of a pxp matrix.}

\item{LT}{Matrix indicating lower triangle of a pxp matrix.}
}
\value{
Updated value for Delta.
}
\description{
Runs one step of the Douglas Rachford algorithm for updating Delta.
}
