% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FB.R
\name{FB}
\alias{FB}
\title{Forward backward algorithm}
\usage{
FB(xi, fb_param1, fb_param2)
}
\arguments{
\item{xi}{Previous value to be updated.}

\item{fb_param1}{First parameter equal to 2 gamma_fb colSums(S Delta xi)}

\item{fb_param2}{Second parameter equal to 4 c gamma_fb}
}
\value{
Updated value for xi.
}
\description{
Runs one step of the forward backward algorithm for updating xi.
}
