% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_fun.R
\name{obj_fun}
\alias{obj_fun}
\title{PCGLASSO objective function}
\usage{
obj_fun(Delta, xi, S, rho, c)
}
\arguments{
\item{Delta}{Partial correlation matrix Delta.}

\item{xi}{Vector of square root diagonals xi.}

\item{S}{Sample covariance matrix.}

\item{rho}{Penalty parameter.}

\item{c}{Diagonal parameter (double usual parameter value).}
}
\value{
Value of objective function evaluated at (Delta, xi).
}
\description{
Evaluates the PCGLASSO objective function (negative penalised likelihood).
}
