% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCGLASSO.R
\name{pcglasso}
\alias{pcglasso}
\title{PCGLASSO}
\usage{
pcglasso(
  S,
  rho,
  c = NULL,
  Theta_start = NULL,
  threshold = 10^(-5),
  max_iter = 10000
)
}
\arguments{
\item{S}{Sample covariance matrix.}

\item{rho}{Penalty parameter.}

\item{c}{Diagonal parameter. Default is the largest possible value for which
the solution exists.}

\item{Theta_start}{Starting value of Theta. Default is the inverse of S
(generalised inverse if S is not positive definite).}

\item{threshold}{Threshold for stopping rule of algorithm.}

\item{max_iter}{Maximum number of iterations.}
}
\value{
Numerical solution Theta to the PCGLASSO optimisation problem.
}
\description{
Numerical solution to the PCGLASSO.
}
