% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Shrink operator}
\usage{
shrink(a, b)
}
\arguments{
\item{a}{Value or vector of number(s) to be shrunk.}

\item{b}{Amount to be shrunk by.}
}
\value{
Value or vector of shrunk number(s).
}
\description{
Performs the shrink operation.
}
