% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_rule.R
\name{stopping_rule}
\alias{stopping_rule}
\title{Stopping rule}
\usage{
stopping_rule(Delta, Delta_old, xi, xi_old, obj, obj_old, threshold)
}
\arguments{
\item{Delta}{New approximated Delta.}

\item{Delta_old}{Previous approximated Delta.}

\item{xi}{New approximated xi.}

\item{xi_old}{Previous approximated xi.}

\item{obj}{New objective function value.}

\item{obj_old}{Previous objective function value.}

\item{threshold}{Threshold for stopping rule to be achieved.}
}
\value{
Indicator of stopping rule.
}
\description{
Evaluates a stopping rule for a Delta, xi update.
}
